Rename Moonbeam datasource processor
The datasource processor seems to apply to Frontier so its better to have the general name.

Also re-exports Frontier as Moonbeam for backwards compatibility.
